Refactor code in `src/` folder
