+++ .github/workflows/ci.yml
@@ -33,3 +33,6 @@ jobs:
 
       - name: Lint with ruff
         run: uv run ruff check .
+
+      - name: Run tests
+        run: uv run pytest tests/ -v
