+++ dev.yml
@@ -7,6 +7,7 @@ up:
 commands:
   generate: python run.py
   server: cd docs && python -m http.server $PORT
+  test: pytest .
 
 check:
   format: ruff check --fix --unsafe-fixes . && ruff format .
