+++ docs/v1/feed.xml
@@ -4,7 +4,7 @@
     <title>AI Model Deprecations</title>
     <link>https://deprecations.info/</link>
     <description>RSS feed tracking deprecations across major AI providers</description>
-    <lastBuildDate>Tue, 26 Aug 2025 23:11:21 GMT</lastBuildDate>
+    <lastBuildDate>Wed, 27 Aug 2025 00:41:23 GMT</lastBuildDate>
     <item>
       <title>OpenAI: Assistants API</title>
       <link>https://platform.openai.com/docs/deprecations#2025-08-20-assistants-api</link>
