+++ pyproject.toml
@@ -19,6 +19,7 @@ dependencies = [
     "ipython>=9.4.0",
     "anthropic>=0.64.0",
     "instructor>=1.10.0",
+    "pytest>=8.0.0",
 ]
 
 [build-system]
@@ -26,6 +27,5 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["."]
-only-include = ["*.py"]
+packages = ["src"]
