+++ src/scrapers/aws_bedrock_scraper.py
@@ -3,8 +3,8 @@
 from typing import List
 from bs4 import BeautifulSoup
 
-from base_scraper import EnhancedBaseScraper
-from models import DeprecationItem
+from ..base_scraper import EnhancedBaseScraper
+from ..models import DeprecationItem
 
 
 class AWSBedrockScraper(EnhancedBaseScraper):
