+++ src/scrapers/cohere_scraper.py
@@ -4,9 +4,9 @@ import re
 from typing import List
 from bs4 import BeautifulSoup
 
-from base_scraper import EnhancedBaseScraper
-from models import DeprecationItem
-from llm_analyzer import LLMAnalyzer
+from ..base_scraper import EnhancedBaseScraper
+from ..models import DeprecationItem
+from ..llm_analyzer import LLMAnalyzer
 
 
 class CohereScraper(EnhancedBaseScraper):
