--- test_basic.py
@@ -1,17 +1,0 @@
-"""Basic smoke tests for the deprecations RSS project."""
-
-
-def test_imports():
-    """Test that main modules can be imported."""
-    import main
-    import providers
-    import rss_gen
-
-    assert main is not None
-    assert providers is not None
-    assert rss_gen is not None
-
-
-def test_basic_functionality():
-    """Placeholder test - to be expanded with real tests."""
-    assert True
